+++ src/Ares/RuleFactory.php
@@ -15,6 +15,7 @@ use Ares\Exception\UnknownValidationRuleIdException;
 use Ares\Rule\AllowedRule;
 use Ares\Rule\BlankableRule;
 use Ares\Rule\DateTimeRule;
+use Ares\Rule\DirectoryRule;
 use Ares\Rule\EmailRule;
 use Ares\Rule\FileRule;
 use Ares\Rule\ForbiddenRule;
@@ -40,6 +41,7 @@ class RuleFactory
         AllowedRule::ID   => ['className' => AllowedRule::class,   'reserved' => false],
         BlankableRule::ID => ['className' => BlankableRule::class, 'reserved' => true],
         DateTimeRule::ID  => ['className' => DateTimeRule::class,  'reserved' => false],
+        DirectoryRule::ID => ['className' => DirectoryRule::class, 'reserved' => false],
         EmailRule::ID     => ['className' => EmailRule::class,     'reserved' => false],
         FileRule::ID      => ['className' => FileRule::class,      'reserved' => false],
         ForbiddenRule::ID => ['className' => ForbiddenRule::class, 'reserved' => false],